Migrate /stats querying to new format, delete old file
